requirements.txt
python-dotenv
pvporcupine
pvrecorder
numpy
tqdm
more-itertools
transformers>=4.19.0
ffmpeg-python==0.2.0
click
pyaudio
SpeechRecognition
pydub
git+https://github.com/openai/whisper.git
--extra-index-url https://download.pytorch.org/whl/cu113
torch
pynput
rich
langchain
openai
fairseq
g2p_en
beepy